Moved all files out of src folder.
